mavlink: move GPS2_RAW to separate stream header
