+++ src/modules/mavlink/mavlink_messages.cpp
@@ -151,6 +151,7 @@ using matrix::wrap_2pi;
 # include "streams/GIMBAL_DEVICE_SET_ATTITUDE.hpp"
 # include "streams/GIMBAL_MANAGER_INFORMATION.hpp"
 # include "streams/GIMBAL_MANAGER_STATUS.hpp"
+# include "streams/GPS2_RAW.hpp"
 # include "streams/LINK_NODE_STATUS.hpp"
 # include "streams/NAMED_VALUE_FLOAT.hpp"
 # include "streams/ODOMETRY.hpp"
@@ -243,18 +244,6 @@ static_assert(MAV_SENSOR_ROTATION_ROLL_90_PITCH_315 == static_cast<MAV_SENSOR_OR
 	      "Roll: 90, Pitch: 315");
 
 static_assert(41 == ROTATION_MAX, "Keep MAV_SENSOR_ROTATION and PX4 Rotation in sync");
-
-static uint16_t cm_uint16_from_m_float(float m)
-{
-	if (m < 0.0f) {
-		return 0;
-
-	} else if (m > 655.35f) {
-		return 65535;
-	}
-
-	return (uint16_t)(m * 100.0f);
-}
 
 void get_mavlink_navigation_mode(const struct vehicle_status_s *const status, uint8_t *mavlink_base_mode,
 				 union px4_custom_mode *custom_mode)
@@ -1402,85 +1391,19 @@ protected:
 			msg.h_acc = gps.eph * 1e3f;
 			msg.v_acc = gps.epv * 1e3f;
 			msg.vel_acc = gps.s_variance_m_s * 1e3f;
-			msg.hdg_acc = gps.c_variance_rad * 1e5f / M_DEG_TO_RAD_F;
-			msg.vel = cm_uint16_from_m_float(gps.vel_m_s);
+			msg.hdg_acc = math::degrees(gps.c_variance_rad) * 1e5f; // degE5
+
+			if (PX4_ISFINITE(gps.vel_m_s) && (fabsf(gps.vel_m_s) >= 0.f)) {
+				msg.vel = gps.vel_m_s * 100.f; // cm/s
+
+			} else {
+				msg.vel = UINT16_MAX; // If unknown, set to: UINT16_MAX
+			}
+
 			msg.cog = math::degrees(wrap_2pi(gps.cog_rad)) * 1e2f;
 			msg.satellites_visible = gps.satellites_used;
 
 			mavlink_msg_gps_raw_int_send_struct(_mavlink->get_channel(), &msg);
-
-			return true;
-		}
-
-		return false;
-	}
-};
-
-class MavlinkStreamGPS2Raw : public MavlinkStream
-{
-public:
-	const char *get_name() const override
-	{
-		return MavlinkStreamGPS2Raw::get_name_static();
-	}
-
-	static constexpr const char *get_name_static()
-	{
-		return "GPS2_RAW";
-	}
-
-	static constexpr uint16_t get_id_static()
-	{
-		return MAVLINK_MSG_ID_GPS2_RAW;
-	}
-
-	uint16_t get_id() override
-	{
-		return get_id_static();
-	}
-
-	static MavlinkStream *new_instance(Mavlink *mavlink)
-	{
-		return new MavlinkStreamGPS2Raw(mavlink);
-	}
-
-	unsigned get_size() override
-	{
-		return _gps2_sub.advertised() ? (MAVLINK_MSG_ID_GPS2_RAW_LEN + MAVLINK_NUM_NON_PAYLOAD_BYTES) : 0;
-	}
-
-private:
-	uORB::Subscription _gps2_sub{ORB_ID(sensor_gps), 1};
-
-	/* do not allow top copying this class */
-	MavlinkStreamGPS2Raw(MavlinkStreamGPS2Raw &) = delete;
-	MavlinkStreamGPS2Raw &operator = (const MavlinkStreamGPS2Raw &) = delete;
-
-protected:
-	explicit MavlinkStreamGPS2Raw(Mavlink *mavlink) : MavlinkStream(mavlink)
-	{}
-
-	bool send() override
-	{
-		sensor_gps_s gps;
-
-		if (_gps2_sub.update(&gps)) {
-			mavlink_gps2_raw_t msg = {};
-
-			msg.time_usec = gps.timestamp;
-			msg.fix_type = gps.fix_type;
-			msg.lat = gps.lat;
-			msg.lon = gps.lon;
-			msg.alt = gps.alt;
-			msg.eph = gps.eph * 1e3f;
-			msg.epv = gps.epv * 1e3f;
-			msg.vel = cm_uint16_from_m_float(gps.vel_m_s);
-			msg.cog = math::degrees(wrap_2pi(gps.cog_rad)) * 1e2f;
-			msg.satellites_visible = gps.satellites_used;
-			//msg.dgps_numch = // Number of DGPS satellites
-			//msg.dgps_age = // Age of DGPS info
-
-			mavlink_msg_gps2_raw_send_struct(_mavlink->get_channel(), &msg);
 
 			return true;
 		}
@@ -1702,7 +1625,9 @@ static const StreamListItem streams_list[] = {
 	create_stream_list_item<MavlinkStreamGpsGlobalOrigin>(),
 #endif // GPS_GLOBAL_ORIGIN_HPP
 	create_stream_list_item<MavlinkStreamGPSRawInt>(),
+#if defined(GPS2_RAW_HPP)
 	create_stream_list_item<MavlinkStreamGPS2Raw>(),
+#endif // GPS2_RAW_HPP
 #if defined(SYSTEM_TIME_HPP)
 	create_stream_list_item<MavlinkStreamSystemTime>(),
 #endif // SYSTEM_TIME_HPP
